pyproject.toml
[project]
name = "main-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.0",
    "openai>=2.30.0",
    "playwright>=1.58.0",
    "pyotp>=2.9.0",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "requests>=2.32.0",
]
